Code review feedback - move config to separate file
