+++ 3rdPartyCypress/cypress.json
@@ -3,5 +3,8 @@
   "screenshotOnRunFailure": false,
   "env": {
     "APP_ENV": "local"
-  }
+  },
+  "ignoreTestFiles": [
+    "**/**/config.js"
+  ]
 }
